+++ CHANGELOG.md
@@ -46,6 +46,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The documentation incorrectly stated that the primary metric for the Reading
   Comprehension task is the Exact Match score. This has been corrected to the
   character-level F1-score. This was fixed by @Rijgersberg ✨
+- There were issues with using Pydantic classes in as `response_format` for LiteLLM
+  models running on custom inference servers. We now use dictionaries instead, which
+  fixes the issue.
 
 ## [v16.16.1] - 2026-02-25